Extract IDependenciesGraphChangeTracker
This pulls out all code from DependenciesGraphProvider that updates the
graph in response to project updates.
